Moved monthly statistics repository methods into its own repository following class responsibility principles
