+++ src/Altinn.Broker.Persistence/DependencyInjection.cs
@@ -1,4 +1,4 @@
-﻿using Altinn.Broker.Core.Repositories;
+using Altinn.Broker.Core.Repositories;
 using Altinn.Broker.Persistence.Helpers;
 using Altinn.Broker.Persistence.Options;
 using Altinn.Broker.Persistence.Repositories;
@@ -20,6 +20,7 @@ public static class DependencyInjection
         services.AddSingleton<ExecuteDBCommandWithRetries>();
         services.AddSingleton<IActorRepository, ActorRepository>();
         services.AddSingleton<IFileTransferRepository, FileTransferRepository>();
+        services.AddSingleton<IMonthlyStatisticsRepository, MonthlyStatisticsRepository>();
         services.AddSingleton<IFileTransferStatusRepository, FileTransferStatusRepository>();
         services.AddSingleton<IActorFileTransferStatusRepository, ActorFileTransferStatusRepository>();
         services.AddSingleton<IServiceOwnerRepository, ServiceOwnerRepository>();
